+++ Tests/Batch1/SimpleTypes/DateTimeTests.cs
@@ -9,11 +9,29 @@ namespace Bridge.ClientTest.SimpleTypes
     public class JsDateTimeTests
     {
         [Test]
-        public void TypePropertiesAreCorrect()
+        public void TypePropertiesAreCorrect_SPI_1607_1608_1609()
         {
             Assert.AreEqual("System.DateTime", typeof(DateTime).FullName, "#2064");
-            object o = new DateTime();
-            Assert.True(o is DateTime, "o is DateTime");
+            Assert.False(typeof(DateTime).IsClass);
+            // #1607 #1608 #1609
+            Assert.True(typeof(IComparable<DateTime>).IsAssignableFrom(typeof(DateTime)));
+            Assert.True(typeof(IEquatable<DateTime>).IsAssignableFrom(typeof(DateTime)));
+            Assert.True(typeof(IFormattable).IsAssignableFrom(typeof(DateTime)));
+
+            object d = new DateTime();
+            Assert.True(d is DateTime);
+            // #1609
+            Assert.True(d is IComparable<DateTime>);
+            // #1608
+            Assert.True(d is IEquatable<DateTime>);
+
+            Assert.True(d is IFormattable);
+
+            var interfaces = typeof(DateTime).GetInterfaces();
+            Assert.AreEqual(4, interfaces.Length);
+            Assert.True(interfaces.Contains(typeof(IComparable<DateTime>)));
+            Assert.True(interfaces.Contains(typeof(IEquatable<DateTime>)));
+            Assert.True(interfaces.Contains(typeof(IFormattable)));
         }
 
         private void AssertDate(DateTime dt, int? year = null, int? month = null, int? day = null, int? hour = null, int? minute = null, int? second = null, int? ms = null)
@@ -77,6 +95,13 @@ namespace Bridge.ClientTest.SimpleTypes
             var dt = Activator.CreateInstance<DateTime>();
             // #1606
             AssertDate(dt, 1, 1, 1);
+        }
+
+        [Test]
+        public void MillisecondSinceEpochConstructorWorks()
+        {
+            var dt = new DateTime(1440L * 60 * 500 * 1000);
+            Assert.AreEqual(1, dt.AddDays(1).Year);
         }
 
         [Test]
@@ -157,14 +182,14 @@ namespace Bridge.ClientTest.SimpleTypes
         }
 
         [Test]
-        public void UTCNowWorks()
+        public void UtcNowWorks()
         {
-            var UTC = DateTime.UtcNow;
+            var utc= DateTime.UtcNow;
             var local = DateTime.Now;
             Assert.True(
                 Math.Abs(
                     (new DateTime(local.GetUtcFullYear(), local.GetUtcMonth(), local.GetUtcDate(), local.GetUtcHours(), local.GetUtcMinutes(), local.GetUtcSeconds(), local.GetUtcMilliseconds())
-                    - UTC).TotalMinutes
+                    - utc).TotalMinutes
                 ) < 1000);
         }
 
@@ -179,7 +204,7 @@ namespace Bridge.ClientTest.SimpleTypes
             Assert.AreEqual(UTC.Hour, dt.GetUtcHours());
             Assert.AreEqual(UTC.Minute, dt.GetUtcMinutes());
             Assert.AreEqual(UTC.Second, dt.GetUtcSeconds());
-            Assert.AreEqual(UTC.GetMilliseconds(), dt.GetUtcMilliseconds());
+            Assert.AreEqual(UTC.Millisecond, dt.GetUtcMilliseconds());
         }
 
         [Test]
@@ -193,7 +218,7 @@ namespace Bridge.ClientTest.SimpleTypes
             Assert.AreEqual(UTC.Hour, dt.GetUtcHours());
             Assert.AreEqual(UTC.Minute, dt.GetUtcMinutes());
             Assert.AreEqual(UTC.Second, dt.GetUtcSeconds());
-            Assert.AreEqual(UTC.GetMilliseconds(), dt.GetUtcMilliseconds());
+            Assert.AreEqual(UTC.Millisecond, dt.GetUtcMilliseconds());
         }
 
         [Test]
@@ -204,7 +229,7 @@ namespace Bridge.ClientTest.SimpleTypes
             Assert.AreEqual(0, dt.Hour);
             Assert.AreEqual(0, dt.Minute);
             Assert.AreEqual(0, dt.Second);
-            Assert.AreEqual(0, dt.GetMilliseconds());
+            Assert.AreEqual(0, dt.Millisecond);
         }
 
         [Test]
@@ -215,6 +240,20 @@ namespace Bridge.ClientTest.SimpleTypes
         }
 
         [Test]
+        public void ToStringWithFormatWorks()
+        {
+            var dt = new DateTime(2011, 7, 12);
+            Assert.AreEqual("2011-07-12", dt.ToString("yyyy-MM-dd"));
+        }
+
+        [Test]
+        public void ToStringWithFormatAndProviderWorks()
+        {
+            var dt = new DateTime(2011, 7, 12);
+            Assert.AreEqual("2011-07-12", dt.ToString("yyyy-MM-dd", CultureInfo.InvariantCulture));
+        }
+
+        [Test]
         public void IFormattableToStringWorks()
         {
             var dt = new DateTime(2011, 7, 12, 13);
@@ -222,6 +261,7 @@ namespace Bridge.ClientTest.SimpleTypes
             Assert.AreEqual(((IFormattable)dt).ToString("yyyy-MM-dd", CultureInfo.CurrentCulture), "2011-07-12");
         }
 
+        // Not C# API
         [Test]
         public void LocaleFormatWorks()
         {
@@ -275,14 +315,14 @@ namespace Bridge.ClientTest.SimpleTypes
         public void GetMillisecondsWorks()
         {
             var dt = new DateTime(2011, 7, 12, 13, 42, 56, 345);
-            Assert.AreEqual(345, dt.GetMilliseconds());
+            Assert.AreEqual(345, dt.Millisecond);
         }
 
         [Test]
         public void GetDayWorks()
         {
             var dt = new DateTime(2011, 7, 12, 13, 42, 56, 345);
-            Assert.AreEqual(2, dt.GetDay());
+            Assert.AreEqual(2, dt.DayOfWeek);
         }
 
         [Test]
@@ -391,10 +431,22 @@ namespace Bridge.ClientTest.SimpleTypes
         }
 
         [Test]
+        public void ParseExactReturnsNullIfTheInputIsInvalid()
+        {
+            Assert.Throws<FormatException>(() => { var dt = DateTime.ParseExact("X", "yyyy-dd-MM"); });
+        }
+
+        [Test]
         public void ParseExactWithCultureWorks()
         {
             var dt = DateTime.ParseExact("2012-12-08", "yyyy-dd-MM", CultureInfo.InvariantCulture);
             AssertDate(dt, 2012, 8, 12);
+        }
+
+        [Test]
+        public void ParseExactWithCultureReturnsNullIfTheInputIsInvalid()
+        {
+            Assert.Throws<FormatException>(() => { var dt = DateTime.ParseExact("X", "yyyy-dd-MM", CultureInfo.InvariantCulture); });
         }
 
         [Test]
@@ -408,6 +460,12 @@ namespace Bridge.ClientTest.SimpleTypes
         }
 
         [Test]
+        public void ParseExactUtcReturnsNullIfTheInputIsInvalid()
+        {
+            Assert.Throws<FormatException>(() => { var dt = DateTime.ParseExact("X", "yyyy-dd-MM", true); });
+        }
+
+        [Test]
         public void ParseExactUTCWithCultureWorks()
         {
             var dt = DateTime.ParseExact("2012-12-08", "yyyy-dd-MM", CultureInfo.InvariantCulture, true);
@@ -415,6 +473,12 @@ namespace Bridge.ClientTest.SimpleTypes
             Assert.AreEqual(2012, dt.GetUtcFullYear());
             Assert.AreEqual(8, dt.GetUtcMonth());
             Assert.AreEqual(12, dt.GetUtcDate());
+        }
+
+        [Test]
+        public void ParseExactUtcWithCultureReturnsNullIfTheInputIsInvalid()
+        {
+            Assert.Throws<FormatException>(() => { var dt = DateTime.ParseExact("X", "yyyy-dd-MM", CultureInfo.InvariantCulture, true); });
         }
 
         [Test]
@@ -468,6 +532,70 @@ namespace Bridge.ClientTest.SimpleTypes
             Assert.AreEqual(milliseconds, dt.GetUtcMilliseconds());
         }
 
+        // Not C# API
+        //[Test]
+        //public void FromUtcYMDWorks()
+        //{
+        //    AssertDateUtc(DateTime.FromUtc(2011, 7, 12), 2011, 7, 12, 0, 0, 0, 0);
+        //}
+
+        // Not C# API
+        //[Test]
+        //public void FromUtcYMDHWorks()
+        //{
+        //    AssertDateUtc(DateTime.FromUtc(2011, 7, 12, 13), 2011, 7, 12, 13, 0, 0, 0);
+        //}
+
+        // Not C# API
+        //[Test]
+        //public void FromUtcYMDHNWorks()
+        //{
+        //    AssertDateUtc(DateTime.FromUtc(2011, 7, 12, 13, 42), 2011, 7, 12, 13, 42, 0, 0);
+        //}
+
+        // Not C# API
+        //[Test]
+        //public void FromUtcYMDHNSWorks()
+        //{
+        //    AssertDateUtc(DateTime.FromUtc(2011, 7, 12, 13, 42, 56), 2011, 7, 12, 13, 42, 56, 0);
+        //}
+
+        // Not C# API
+        //[Test]
+        //public void FromUtcYMDHNSUWorks()
+        //{
+        //    AssertDateUtc(DateTime.FromUtc(2011, 7, 12, 13, 42, 56, 345), 2011, 7, 12, 13, 42, 56, 345);
+        //}
+
+        [Test]
+        public void UtcYMDWorks()
+        {
+            AssertDateUTC(new DateTime(DateTime.Utc(2011, 7, 12)), 2011, 7, 12, 0, 0, 0, 0);
+        }
+
+        [Test]
+        public void UtcYMDHWorks()
+        {
+            AssertDateUTC(new DateTime(DateTime.Utc(2011, 7, 12, 13)), 2011, 7, 12, 13, 0, 0, 0);
+        }
+
+        [Test]
+        public void UtcYMDHNWorks()
+        {
+            AssertDateUTC(new DateTime(DateTime.Utc(2011, 7, 12, 13, 42)), 2011, 7, 12, 13, 42, 0, 0);
+        }
+
+        [Test]
+        public void UtcYMDHNSWorks()
+        {
+            AssertDateUTC(new DateTime(DateTime.Utc(2011, 7, 12, 13, 42, 56)), 2011, 7, 12, 13, 42, 56, 0);
+        }
+
+        [Test]
+        public void UtcYMDHNSUWorks()
+        {
+            AssertDateUTC(new DateTime(DateTime.Utc(2011, 7, 12, 13, 42, 56, 345)), 2011, 7, 12, 13, 42, 56, 345);
+        }
         [Test]
         public void SubtractingDatesWorks()
         {
@@ -481,6 +609,28 @@ namespace Bridge.ClientTest.SimpleTypes
             Assert.AreDeepEqual(new TimeSpan(1, 0, 0, 0), new DateTime(2011, 6, 12).Subtract(new DateTime(2011, 6, 11)));
             Assert.AreDeepEqual(new TimeSpan(1, 2, 0, 0), new DateTime(2011, 6, 12, 15, 0, 0).Subtract(new DateTime(2011, 6, 11, 13, 0, 0)));
         }
+
+        // Not C# API
+        //[Test]
+        //public void AreEqualWorks()
+        //{
+        //    Assert.True(DateTime.AreEqual(new DateTime(2011, 7, 12), new DateTime(2011, 7, 12)));
+        //    Assert.False(DateTime.AreEqual(new DateTime(2011, 7, 12), new DateTime(2011, 7, 13)));
+        //    Assert.AreStrictEqual(DateTime.AreEqual(new DateTime(2011, 7, 12), (DateTime?)null), false);
+        //    Assert.AreStrictEqual(DateTime.AreEqual((DateTime?)null, new DateTime(2011, 7, 12)), false);
+        //    Assert.AreStrictEqual(DateTime.AreEqual((DateTime?)null, (DateTime?)null), true);
+        //}
+
+        // Not C# API
+        //[Test]
+        //public void AreNotEqualWorks()
+        //{
+        //    Assert.False(DateTime.AreNotEqual(new DateTime(2011, 7, 12), new DateTime(2011, 7, 12)));
+        //    Assert.True(DateTime.AreNotEqual(new DateTime(2011, 7, 12), new DateTime(2011, 7, 13)));
+        //    Assert.AreStrictEqual(DateTime.AreNotEqual(new DateTime(2011, 7, 12), (DateTime?)null), true);
+        //    Assert.AreStrictEqual(DateTime.AreNotEqual((DateTime?)null, new DateTime(2011, 7, 12)), true);
+        //    Assert.AreStrictEqual(DateTime.AreNotEqual((DateTime?)null, (DateTime?)null), false);
+        //}
 
         [Test]
         public void DateEqualityWorks()
@@ -544,147 +694,6 @@ namespace Bridge.ClientTest.SimpleTypes
             Assert.True(new DateTime(2011, 7, 13) >= new DateTime(2011, 7, 12));
         }
 
-        // TODO DateTime.SetFullYear
-        //[Test]
-        //public void SetFullYearWithOneParameterWorks()
-        //{
-        //    var dt = new DateTime(2000, 0, 1);
-        //    dt.SetFullYear(2021);
-        //    Assert.AreEqual(dt.Year, 2021);
-        //}
-
-        //[Test]
-        //public void SetFullYearWithTwoParametersWorks() {
-        //    var dt = new DateTime(2000, 0, 1);
-        //    dt.SetFullYear(2021, 7);
-        //    Assert.AreEqual(dt.Year, 2021);
-        //    Assert.AreEqual(dt.Month, 7);
-        //}
-
-        //[Test]
-        //public void SetFullYearWithThreeParametersWorks() {
-        //    var dt = new DateTime(2000, 0, 1);
-        //    dt.SetFullYear(2021, 7, 13);
-        //    Assert.AreEqual(dt.Year, 2021);
-        //    Assert.AreEqual(dt.Month, 7);
-        //    Assert.AreEqual(dt.Day, 13);
-        //}
-
-        //[Test]
-        //public void SetMonthWorks() {
-        //    var dt = new DateTime(2000, 0, 1);
-        //    dt.SetMonth(3);
-        //    Assert.AreEqual(dt.Month, 3);
-        //}
-
-        //[Test]
-        //public void SetDateWorks() {
-        //    var dt = new DateTime(2000, 0, 1);
-        //    dt.SetDate(12);
-        //    Assert.AreEqual(dt.Day, 12);
-        //}
-
-        //[Test]
-        //public void SetHoursWorks() {
-        //    var dt = new DateTime(2000, 0, 1);
-        //    dt.SetHours(11);
-        //    Assert.AreEqual(dt.Hour, 11);
-        //}
-
-        //[Test]
-        //public void SetMinutesWorks() {
-        //    var dt = new DateTime(2000, 0, 1);
-        //    dt.SetMinutes(34);
-        //    Assert.AreEqual(dt.Minute, 34);
-        //}
-
-        //[Test]
-        //public void SetSecondsWorks() {
-        //    var dt = new DateTime(2000, 0, 1);
-        //    dt.SetSeconds(23);
-        //    Assert.AreEqual(dt.Second, 23);
-        //}
-
-        //[Test]
-        //public void SetMillisecondsWorks() {
-        //    var dt = new DateTime(2000, 0, 1);
-        //    dt.SetMilliseconds(435);
-        //    Assert.AreEqual(dt.GetMilliseconds(), 435);
-        //}
-
-        //[Test]
-        //public void SetTimeWorks() {
-        //    var dt = new DateTime();
-        //    dt.SetTime(3498302349234L);
-        //    Assert.AreEqual(dt.GetTime(), 3498302349234L);
-        //}
-
-        //[Test]
-        //public void SetUTCFullYearWithOneParameterWorks() {
-        //    var dt = new DateTime(DateTime.UTC(2000, 0, 1));
-        //    dt.SetUTCFullYear(2021);
-        //    Assert.AreEqual(dt.GetUTCFullYear(), 2021);
-        //}
-
-        //[Test]
-        //public void SetUTCFullYearWithTwoParametersWorks() {
-        //    var dt = new DateTime(DateTime.UTC(2000, 0, 1));
-        //    dt.SetUTCFullYear(2021, 7);
-        //    Assert.AreEqual(dt.GetUTCFullYear(), 2021);
-        //    Assert.AreEqual(dt.GetUtcMonth(), 7);
-        //}
-
-        //[Test]
-        //public void SetUTCFullYearWithThreeParametersWorks() {
-        //    var dt = new DateTime(DateTime.UTC(2000, 0, 1));
-        //    dt.SetUTCFullYear(2021, 7, 13);
-        //    Assert.AreEqual(dt.GetUTCFullYear(), 2021);
-        //    Assert.AreEqual(dt.GetUtcMonth(), 7);
-        //    Assert.AreEqual(dt.GetUTCDate(), 13);
-        //}
-
-        //[Test]
-        //public void SetUTCMonthWorks() {
-        //    var dt = new DateTime(DateTime.UTC(2000, 0, 1));
-        //    dt.SetUTCMonth(3);
-        //    Assert.AreEqual(dt.GetUtcMonth(), 3);
-        //}
-
-        //[Test]
-        //public void SetUTCDateWorks() {
-        //    var dt = new DateTime(DateTime.UTC(2000, 0, 1));
-        //    dt.SetUTCDate(12);
-        //    Assert.AreEqual(dt.GetUTCDate(), 12);
-        //}
-
-        //[Test]
-        //public void SetUTCHoursWorks() {
-        //    var dt = new DateTime(DateTime.UTC(2000, 0, 1));
-        //    dt.SetUTCHours(11);
-        //    Assert.AreEqual(dt.GetUTCHours(), 11);
-        //}
-
-        //[Test]
-        //public void SetUTCMinutesWorks() {
-        //    var dt = new DateTime(DateTime.UTC(2000, 0, 1));
-        //    dt.SetUTCMinutes(34);
-        //    Assert.AreEqual(dt.GetUTCMinutes(), 34);
-        //}
-
-        //[Test]
-        //public void SetUTCSecondsWorks() {
-        //    var dt = new DateTime(DateTime.UTC(2000, 0, 1));
-        //    dt.SetUTCSeconds(23);
-        //    Assert.AreEqual(dt.GetUTCSeconds(), 23);
-        //}
-
-        //[Test]
-        //public void SetUTCMillisecondsWorks() {
-        //    var dt = new DateTime(DateTime.UTC(2000, 0, 1));
-        //    dt.SetUTCMilliseconds(435);
-        //    Assert.AreEqual(dt.GetUTCMilliseconds(), 435);
-        //}
-
         [Test]
         public void GetHashCodeWorks()
         {
@@ -704,22 +713,26 @@ namespace Bridge.ClientTest.SimpleTypes
         }
 
         [Test]
-        public void DateTimeEqualsWorks()
+        public void IEquatableEqualsWorks_SPI_1608()
         {
             Assert.True(new DateTime(0).Equals(new DateTime(0)));
             Assert.False(new DateTime(10000).Equals(new DateTime(0)));
             Assert.False(new DateTime(0).Equals(new DateTime(10000)));
             Assert.True(new DateTime(10000).Equals(new DateTime(10000)));
-        }
 
-        [IgnoreTest(Until = Constants.IGNORE_DATE)]
-        [Test]
-        public void IEquatableEqualsWorks()
-        {
             Assert.True(((IEquatable<DateTime>)new DateTime(0)).Equals(new DateTime(0)));
             Assert.False(((IEquatable<DateTime>)new DateTime(10000)).Equals(new DateTime(0)));
             Assert.False(((IEquatable<DateTime>)new DateTime(0)).Equals(new DateTime(10000)));
             Assert.True(((IEquatable<DateTime>)new DateTime(10000)).Equals(new DateTime(10000)));
+        }
+
+        [Test]
+        public void StaticEqualsWorks()
+        {
+            Assert.True(DateTime.Equals(new DateTime(0), new DateTime(0)));
+            Assert.False(DateTime.Equals(new DateTime(10000), new DateTime(0)));
+            Assert.False(DateTime.Equals(new DateTime(0), new DateTime(10000)));
+            Assert.True(DateTime.Equals(new DateTime(10000), new DateTime(10000)));
         }
 
         [Test]
@@ -730,13 +743,181 @@ namespace Bridge.ClientTest.SimpleTypes
             Assert.True(new DateTime(0).CompareTo(new DateTime(10000)) < 0);
         }
 
-        [IgnoreTest(Until = Constants.IGNORE_DATE)]
         [Test]
-        public void IComparableCompareToWorks()
+        public void StaticCompareWorks()
         {
+            Assert.True(DateTime.Compare(new DateTime(0), new DateTime(0)) == 0);
+            Assert.True(DateTime.Compare(new DateTime(10000), new DateTime(0)) > 0);
+            Assert.True(DateTime.Compare(new DateTime(0), new DateTime(10000)) < 0);
+        }
+
+        [Test]
+        public void IComparableCompareToWorks_SPI_1609()
+        {
+            // #1609
             Assert.True(((IComparable<DateTime>)new DateTime(0)).CompareTo(new DateTime(0)) == 0);
             Assert.True(((IComparable<DateTime>)new DateTime(10000)).CompareTo(new DateTime(0)) > 0);
             Assert.True(((IComparable<DateTime>)new DateTime(0)).CompareTo(new DateTime(10000)) < 0);
+        }
+
+        [Test]
+        public void DatePropertyWorks()
+        {
+            var dt = new DateTime(2012, 8, 12, 13, 14, 15, 16);
+            Assert.AreEqual(new DateTime(2012, 8, 12), dt.Date);
+        }
+
+        [Test]
+        public void DayPropertyWorks()
+        {
+            var dt = new DateTime(2011, 7, 12, 13, 42, 56, 345);
+            Assert.AreEqual(12, dt.Day);
+        }
+
+        [Test]
+        public void DayOfWeekPropertyWorks()
+        {
+            var dt = new DateTime(2011, 8, 12, 13, 42, 56, 345);
+            Assert.AreEqual(DayOfWeek.Friday, dt.DayOfWeek);
+        }
+
+        [Test]
+        public void DayOfYearPropertyWorks()
+        {
+            var dt = new DateTime(2011, 7, 12, 13, 42, 56, 345);
+            Assert.AreEqual(193, dt.DayOfYear);
+        }
+
+        [Test]
+        public void HourPropertyWorks()
+        {
+            var dt = new DateTime(2011, 7, 12, 13, 42, 56, 345);
+            Assert.AreEqual(13, dt.Hour);
+        }
+
+        [Test]
+        public void MillisecondPropertyWorks()
+        {
+            var dt = new DateTime(2011, 7, 12, 13, 42, 56, 345);
+            Assert.AreEqual(345, dt.Millisecond);
+        }
+
+        [Test]
+        public void MinutePropertyWorks()
+        {
+            var dt = new DateTime(2011, 7, 12, 13, 42, 56, 345);
+            Assert.AreEqual(42, dt.Minute);
+        }
+
+        [Test]
+        public void MonthPropertyWorks()
+        {
+            var dt = new DateTime(2011, 7, 12, 13, 42, 56, 345);
+            Assert.AreEqual(7, dt.Month);
+        }
+
+        [Test]
+        public void SecondPropertyWorks()
+        {
+            var dt = new DateTime(2011, 7, 12, 13, 42, 56, 345);
+            Assert.AreEqual(56, dt.Second);
+        }
+
+        [Test]
+        public void YearPropertyWorks()
+        {
+            var dt = new DateTime(2011, 7, 12, 13, 42, 56, 345);
+            Assert.AreEqual(2011, dt.Year);
+        }
+
+        [Test]
+        public void AddDaysWorks()
+        {
+            var dt = new DateTime(2011, 7, 12, 2, 42, 56, 345);
+            var actual = dt.AddDays(2.5);
+            Assert.AreEqual(new DateTime(2011, 7, 14, 14, 42, 56, 345), actual);
+            Assert.AreEqual(new DateTime(2011, 7, 12, 2, 42, 56, 345), dt);
+        }
+
+        [Test]
+        public void AddHoursWorks()
+        {
+            var dt = new DateTime(2011, 7, 12, 2, 42, 56, 345);
+            var actual = dt.AddHours(2.5);
+            Assert.AreEqual(new DateTime(2011, 7, 12, 5, 12, 56, 345), actual);
+            Assert.AreEqual(new DateTime(2011, 7, 12, 2, 42, 56, 345), dt);
+        }
+
+        [Test]
+        public void AddMillisecondsWorks()
+        {
+            var dt = new DateTime(2011, 7, 12, 2, 42, 56, 345);
+            var actual = dt.AddMilliseconds(250.4);
+            Assert.AreEqual(new DateTime(2011, 7, 12, 2, 42, 56, 595), actual);
+            Assert.AreEqual(new DateTime(2011, 7, 12, 2, 42, 56, 345), dt);
+        }
+
+        [Test]
+        public void AddMinutesWorks()
+        {
+            var dt = new DateTime(2011, 7, 12, 2, 42, 56, 345);
+            var actual = dt.AddMinutes(2.5);
+            Assert.AreEqual(new DateTime(2011, 7, 12, 2, 45, 26, 345), actual);
+            Assert.AreEqual(new DateTime(2011, 7, 12, 2, 42, 56, 345), dt);
+        }
+
+        [Test]
+        public void AddMonthsWorks()
+        {
+            var dt = new DateTime(2011, 7, 12, 2, 42, 56, 345);
+            var actual = dt.AddMonths(6);
+            Assert.AreEqual(new DateTime(2012, 1, 12, 2, 42, 56, 345), actual);
+            Assert.AreEqual(new DateTime(2011, 7, 12, 2, 42, 56, 345), dt);
+        }
+
+        [Test]
+        public void AddSecondsWorks()
+        {
+            var dt = new DateTime(2011, 7, 12, 2, 42, 56, 345);
+            var actual = dt.AddSeconds(2.5);
+            Assert.AreEqual(new DateTime(2011, 7, 12, 2, 42, 58, 845), actual);
+            Assert.AreEqual(new DateTime(2011, 7, 12, 2, 42, 56, 345), dt);
+        }
+
+        [Test]
+        public void AddYearsWorks()
+        {
+            var dt = new DateTime(2011, 7, 12, 2, 42, 56, 345);
+            var actual = dt.AddYears(3);
+            Assert.AreEqual(new DateTime(2014, 7, 12, 2, 42, 56, 345), actual);
+            Assert.AreEqual(new DateTime(2011, 7, 12, 2, 42, 56, 345), dt);
+        }
+
+        [Test]
+        public void DaysInMonthWorks()
+        {
+            Assert.AreEqual(31, DateTime.DaysInMonth(2013, 1));
+            Assert.AreEqual(28, DateTime.DaysInMonth(2013, 2));
+            Assert.AreEqual(31, DateTime.DaysInMonth(2013, 3));
+            Assert.AreEqual(30, DateTime.DaysInMonth(2013, 4));
+            Assert.AreEqual(31, DateTime.DaysInMonth(2013, 5));
+            Assert.AreEqual(30, DateTime.DaysInMonth(2013, 6));
+            Assert.AreEqual(31, DateTime.DaysInMonth(2013, 7));
+            Assert.AreEqual(31, DateTime.DaysInMonth(2013, 8));
+            Assert.AreEqual(30, DateTime.DaysInMonth(2013, 9));
+            Assert.AreEqual(31, DateTime.DaysInMonth(2013, 10));
+            Assert.AreEqual(30, DateTime.DaysInMonth(2013, 11));
+            Assert.AreEqual(31, DateTime.DaysInMonth(2013, 12));
+            Assert.AreEqual(28, DateTime.DaysInMonth(2003, 2));
+            Assert.AreEqual(29, DateTime.DaysInMonth(2004, 2));
+        }
+
+        [Test]
+        public void IsLeapYearWorks()
+        {
+            Assert.True(DateTime.IsLeapYear(2004));
+            Assert.True(DateTime.IsLeapYear(2000));
+            Assert.False(DateTime.IsLeapYear(2003));
         }
 
         [Test(ExpectedCount = 1)]
