+++ Tests/Runner/Batch4/Bridge.Test/bridge.clienttest.batch4.runner.js
@@ -1229,99 +1229,6 @@ Bridge.assembly("Bridge.Test.Bridge.ClientTest.Batch4", function ($asm, globals)
             QUnit.test("DateTests - SetUtcMillisecondsWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTests.SetUtcMillisecondsWorks);
             QUnit.test("DateTests - GetHashCodeWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTests.GetHashCodeWorks);
             QUnit.test("DateTests - EqualsWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTests.EqualsWorks);
-            QUnit.test("DateTimeTests - TypePropertiesAreCorrect_SPI_1607_1608_1609", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests.TypePropertiesAreCorrect_SPI_1607_1608_1609);
-            QUnit.test("DateTimeTests - DefaultConstructorWorks_SPI_1606", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests.DefaultConstructorWorks_SPI_1606);
-            QUnit.test("DateTimeTests - DefaultValueWorks_SPI_1606", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests.DefaultValueWorks_SPI_1606);
-            QUnit.test("DateTimeTests - CreatingInstanceReturnsDateWithZeroValue_SPI_1606", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests.CreatingInstanceReturnsDateWithZeroValue_SPI_1606);
-            QUnit.test("DateTimeTests - MillisecondSinceEpochConstructorWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests.MillisecondSinceEpochConstructorWorks);
-            QUnit.test("DateTimeTests - StringConstructorWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests.StringConstructorWorks);
-            QUnit.test("DateTimeTests - YMDConstructorWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests.YMDConstructorWorks);
-            QUnit.test("DateTimeTests - YMDHConstructorWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests.YMDHConstructorWorks);
-            QUnit.test("DateTimeTests - YMDHNConstructorWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests.YMDHNConstructorWorks);
-            QUnit.test("DateTimeTests - YMDHNSConstructorWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests.YMDHNSConstructorWorks);
-            QUnit.test("DateTimeTests - YMDHNSUConstructorWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests.YMDHNSUConstructorWorks);
-            QUnit.test("DateTimeTests - NowWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests.NowWorks);
-            QUnit.test("DateTimeTests - UtcNowWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests.UtcNowWorks);
-            QUnit.test("DateTimeTests - ToUniversalWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests.ToUniversalWorks);
-            QUnit.test("DateTimeTests - ToLocalWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests.ToLocalWorks);
-            QUnit.test("DateTimeTests - TodayWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests.TodayWorks);
-            QUnit.test("DateTimeTests - FormatWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests.FormatWorks);
-            QUnit.test("DateTimeTests - ToStringWithFormatWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests.ToStringWithFormatWorks);
-            QUnit.test("DateTimeTests - ToStringWithFormatAndProviderWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests.ToStringWithFormatAndProviderWorks);
-            QUnit.test("DateTimeTests - IFormattableToStringWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests.IFormattableToStringWorks);
-            QUnit.test("DateTimeTests - GetFullYearWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests.GetFullYearWorks);
-            QUnit.test("DateTimeTests - GetMonthWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests.GetMonthWorks);
-            QUnit.test("DateTimeTests - GetDateWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests.GetDateWorks);
-            QUnit.test("DateTimeTests - GetHoursWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests.GetHoursWorks);
-            QUnit.test("DateTimeTests - GetMinutesWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests.GetMinutesWorks);
-            QUnit.test("DateTimeTests - GetSecondsWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests.GetSecondsWorks);
-            QUnit.test("DateTimeTests - GetMillisecondsWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests.GetMillisecondsWorks);
-            QUnit.test("DateTimeTests - GetDayWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests.GetDayWorks);
-            QUnit.test("DateTimeTests - GetTimeWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests.GetTimeWorks);
-            QUnit.test("DateTimeTests - ValueOfWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests.ValueOfWorks);
-            QUnit.test("DateTimeTests - GetUtcFullYearWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests.GetUtcFullYearWorks);
-            QUnit.test("DateTimeTests - GetUtcMonthWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests.GetUtcMonthWorks);
-            QUnit.test("DateTimeTests - GetUtcDateWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests.GetUtcDateWorks);
-            QUnit.test("DateTimeTests - GetUtcHoursWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests.GetUtcHoursWorks);
-            QUnit.test("DateTimeTests - GetUtcMinutesWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests.GetUtcMinutesWorks);
-            QUnit.test("DateTimeTests - GetUtcSecondsWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests.GetUtcSecondsWorks);
-            QUnit.test("DateTimeTests - GetUtcMillisecondsWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests.GetUtcMillisecondsWorks);
-            QUnit.test("DateTimeTests - GetUtcDayWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests.GetUtcDayWorks);
-            QUnit.test("DateTimeTests - ParseWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests.ParseWorks);
-            QUnit.test("DateTimeTests - ParseExactWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests.ParseExactWorks);
-            QUnit.test("DateTimeTests - ParseExactReturnsNullIfTheInputIsInvalid", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests.ParseExactReturnsNullIfTheInputIsInvalid);
-            QUnit.test("DateTimeTests - ParseExactWithCultureWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests.ParseExactWithCultureWorks);
-            QUnit.test("DateTimeTests - ParseExactWithCultureReturnsNullIfTheInputIsInvalid", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests.ParseExactWithCultureReturnsNullIfTheInputIsInvalid);
-            QUnit.test("DateTimeTests - ParseExactUtcWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests.ParseExactUtcWorks);
-            QUnit.test("DateTimeTests - ParseExactUtcReturnsNullIfTheInputIsInvalid", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests.ParseExactUtcReturnsNullIfTheInputIsInvalid);
-            QUnit.test("DateTimeTests - ParseExactUtcWithCultureWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests.ParseExactUtcWithCultureWorks);
-            QUnit.test("DateTimeTests - ParseExactUtcWithCultureReturnsNullIfTheInputIsInvalid", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests.ParseExactUtcWithCultureReturnsNullIfTheInputIsInvalid);
-            QUnit.test("DateTimeTests - ToDateStringWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests.ToDateStringWorks);
-            QUnit.test("DateTimeTests - ToTimeStringWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests.ToTimeStringWorks);
-            QUnit.test("DateTimeTests - ToUtcStringWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests.ToUtcStringWorks);
-            QUnit.test("DateTimeTests - ToLocaleDateStringWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests.ToLocaleDateStringWorks);
-            QUnit.test("DateTimeTests - ToLocaleTimeStringWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests.ToLocaleTimeStringWorks);
-            QUnit.test("DateTimeTests - UtcYMDWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests.UtcYMDWorks);
-            QUnit.test("DateTimeTests - UtcYMDHWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests.UtcYMDHWorks);
-            QUnit.test("DateTimeTests - UtcYMDHNWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests.UtcYMDHNWorks);
-            QUnit.test("DateTimeTests - UtcYMDHNSWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests.UtcYMDHNSWorks);
-            QUnit.test("DateTimeTests - UtcYMDHNSUWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests.UtcYMDHNSUWorks);
-            QUnit.test("DateTimeTests - SubtractingDatesWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests.SubtractingDatesWorks);
-            QUnit.test("DateTimeTests - SubtractMethodReturningTimeSpanWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests.SubtractMethodReturningTimeSpanWorks);
-            QUnit.test("DateTimeTests - DateEqualityWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests.DateEqualityWorks);
-            QUnit.test("DateTimeTests - DateInequalityWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests.DateInequalityWorks);
-            QUnit.test("DateTimeTests - DateLessThanWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests.DateLessThanWorks);
-            QUnit.test("DateTimeTests - DateLessEqualWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests.DateLessEqualWorks);
-            QUnit.test("DateTimeTests - DateGreaterThanWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests.DateGreaterThanWorks);
-            QUnit.test("DateTimeTests - DateGreaterEqualWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests.DateGreaterEqualWorks);
-            QUnit.test("DateTimeTests - ConvertingDateToMutableDateReturnsANewButEqualInstance", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests.ConvertingDateToMutableDateReturnsANewButEqualInstance);
-            QUnit.test("DateTimeTests - ConvertingMutableDateToDateReturnsANewButEqualInstance", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests.ConvertingMutableDateToDateReturnsANewButEqualInstance);
-            QUnit.test("DateTimeTests - GetHashCodeWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests.GetHashCodeWorks);
-            QUnit.test("DateTimeTests - EqualsWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests.EqualsWorks);
-            QUnit.test("DateTimeTests - IEquatableEqualsWorks_SPI_1608", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests.IEquatableEqualsWorks_SPI_1608);
-            QUnit.test("DateTimeTests - StaticEqualsWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests.StaticEqualsWorks);
-            QUnit.test("DateTimeTests - CompareToWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests.CompareToWorks);
-            QUnit.test("DateTimeTests - StaticCompareWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests.StaticCompareWorks);
-            QUnit.test("DateTimeTests - IComparableCompareToWorks_SPI_1609", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests.IComparableCompareToWorks_SPI_1609);
-            QUnit.test("DateTimeTests - DatePropertyWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests.DatePropertyWorks);
-            QUnit.test("DateTimeTests - DayPropertyWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests.DayPropertyWorks);
-            QUnit.test("DateTimeTests - DayOfWeekPropertyWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests.DayOfWeekPropertyWorks);
-            QUnit.test("DateTimeTests - DayOfYearPropertyWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests.DayOfYearPropertyWorks);
-            QUnit.test("DateTimeTests - HourPropertyWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests.HourPropertyWorks);
-            QUnit.test("DateTimeTests - MillisecondPropertyWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests.MillisecondPropertyWorks);
-            QUnit.test("DateTimeTests - MinutePropertyWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests.MinutePropertyWorks);
-            QUnit.test("DateTimeTests - MonthPropertyWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests.MonthPropertyWorks);
-            QUnit.test("DateTimeTests - SecondPropertyWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests.SecondPropertyWorks);
-            QUnit.test("DateTimeTests - YearPropertyWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests.YearPropertyWorks);
-            QUnit.test("DateTimeTests - AddDaysWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests.AddDaysWorks);
-            QUnit.test("DateTimeTests - AddHoursWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests.AddHoursWorks);
-            QUnit.test("DateTimeTests - AddMillisecondsWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests.AddMillisecondsWorks);
-            QUnit.test("DateTimeTests - AddMinutesWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests.AddMinutesWorks);
-            QUnit.test("DateTimeTests - AddMonthsWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests.AddMonthsWorks);
-            QUnit.test("DateTimeTests - AddSecondsWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests.AddSecondsWorks);
-            QUnit.test("DateTimeTests - AddYearsWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests.AddYearsWorks);
-            QUnit.test("DateTimeTests - DaysInMonthWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests.DaysInMonthWorks);
-            QUnit.test("DateTimeTests - IsLeapYearWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests.IsLeapYearWorks);
             QUnit.test("DecimalTests - TypePropertiesAreCorrect_SPI_1717", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DecimalTests.TypePropertiesAreCorrect_SPI_1717);
             QUnit.test("DecimalTests - DefaultValueIsDecimal0", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DecimalTests.DefaultValueIsDecimal0);
             QUnit.test("DecimalTests - CreatingInstanceReturnsZero", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DecimalTests.CreatingInstanceReturnsZero);
@@ -4140,491 +4047,6 @@ Bridge.assembly("Bridge.Test.Bridge.ClientTest.Batch4", function ($asm, globals)
                 var $t;
                 if (this.context == null) {
                     this.context = ($t = new Bridge.Test.Runtime.FixtureContext(), $t.Project = "Batch4", $t.ClassName = "Bridge.ClientTest.Batch4.DateTimeFormatInfoTests", $t.File = "Batch4\\DateTimeFormatInfoTests.cs", $t);
-                }
-                return this.context;
-            }
-        }
-    });
-
-    Bridge.define("Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests", {
-        inherits: [Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests)],
-        statics: {
-            methods: {
-                TypePropertiesAreCorrect_SPI_1607_1608_1609: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "TypePropertiesAreCorrect_SPI_1607_1608_1609()", $t.Line = "11", $t));
-                    t.Fixture.TypePropertiesAreCorrect_SPI_1607_1608_1609();
-                },
-                DefaultConstructorWorks_SPI_1606: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "DefaultConstructorWorks_SPI_1606()", $t.Line = "37", $t));
-                    t.Fixture.DefaultConstructorWorks_SPI_1606();
-                },
-                DefaultValueWorks_SPI_1606: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "DefaultValueWorks_SPI_1606()", $t.Line = "45", $t));
-                    t.Fixture.DefaultValueWorks_SPI_1606();
-                },
-                CreatingInstanceReturnsDateWithZeroValue_SPI_1606: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "CreatingInstanceReturnsDateWithZeroValue_SPI_1606()", $t.Line = "53", $t));
-                    t.Fixture.CreatingInstanceReturnsDateWithZeroValue_SPI_1606();
-                },
-                MillisecondSinceEpochConstructorWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "MillisecondSinceEpochConstructorWorks()", $t.Line = "61", $t));
-                    t.Fixture.MillisecondSinceEpochConstructorWorks();
-                },
-                StringConstructorWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "StringConstructorWorks()", $t.Line = "68", $t));
-                    t.Fixture.StringConstructorWorks();
-                },
-                YMDConstructorWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "YMDConstructorWorks()", $t.Line = "77", $t));
-                    t.Fixture.YMDConstructorWorks();
-                },
-                YMDHConstructorWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "YMDHConstructorWorks()", $t.Line = "86", $t));
-                    t.Fixture.YMDHConstructorWorks();
-                },
-                YMDHNConstructorWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "YMDHNConstructorWorks()", $t.Line = "96", $t));
-                    t.Fixture.YMDHNConstructorWorks();
-                },
-                YMDHNSConstructorWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "YMDHNSConstructorWorks()", $t.Line = "107", $t));
-                    t.Fixture.YMDHNSConstructorWorks();
-                },
-                YMDHNSUConstructorWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "YMDHNSUConstructorWorks()", $t.Line = "119", $t));
-                    t.Fixture.YMDHNSUConstructorWorks();
-                },
-                NowWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "NowWorks()", $t.Line = "132", $t));
-                    t.Fixture.NowWorks();
-                },
-                UtcNowWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "UtcNowWorks()", $t.Line = "139", $t));
-                    t.Fixture.UtcNowWorks();
-                },
-                ToUniversalWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "ToUniversalWorks()", $t.Line = "149", $t));
-                    t.Fixture.ToUniversalWorks();
-                },
-                ToLocalWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "ToLocalWorks()", $t.Line = "163", $t));
-                    t.Fixture.ToLocalWorks();
-                },
-                TodayWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "TodayWorks()", $t.Line = "177", $t));
-                    t.Fixture.TodayWorks();
-                },
-                FormatWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "FormatWorks()", $t.Line = "188", $t));
-                    t.Fixture.FormatWorks();
-                },
-                ToStringWithFormatWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "ToStringWithFormatWorks()", $t.Line = "195", $t));
-                    t.Fixture.ToStringWithFormatWorks();
-                },
-                ToStringWithFormatAndProviderWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "ToStringWithFormatAndProviderWorks()", $t.Line = "202", $t));
-                    t.Fixture.ToStringWithFormatAndProviderWorks();
-                },
-                IFormattableToStringWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "IFormattableToStringWorks()", $t.Line = "209", $t));
-                    t.Fixture.IFormattableToStringWorks();
-                },
-                GetFullYearWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "GetFullYearWorks()", $t.Line = "224", $t));
-                    t.Fixture.GetFullYearWorks();
-                },
-                GetMonthWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "GetMonthWorks()", $t.Line = "231", $t));
-                    t.Fixture.GetMonthWorks();
-                },
-                GetDateWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "GetDateWorks()", $t.Line = "238", $t));
-                    t.Fixture.GetDateWorks();
-                },
-                GetHoursWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "GetHoursWorks()", $t.Line = "245", $t));
-                    t.Fixture.GetHoursWorks();
-                },
-                GetMinutesWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "GetMinutesWorks()", $t.Line = "252", $t));
-                    t.Fixture.GetMinutesWorks();
-                },
-                GetSecondsWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "GetSecondsWorks()", $t.Line = "259", $t));
-                    t.Fixture.GetSecondsWorks();
-                },
-                GetMillisecondsWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "GetMillisecondsWorks()", $t.Line = "266", $t));
-                    t.Fixture.GetMillisecondsWorks();
-                },
-                GetDayWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "GetDayWorks()", $t.Line = "273", $t));
-                    t.Fixture.GetDayWorks();
-                },
-                GetTimeWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "GetTimeWorks()", $t.Line = "280", $t));
-                    t.Fixture.GetTimeWorks();
-                },
-                ValueOfWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "ValueOfWorks()", $t.Line = "287", $t));
-                    t.Fixture.ValueOfWorks();
-                },
-                GetUtcFullYearWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "GetUtcFullYearWorks()", $t.Line = "302", $t));
-                    t.Fixture.GetUtcFullYearWorks();
-                },
-                GetUtcMonthWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "GetUtcMonthWorks()", $t.Line = "309", $t));
-                    t.Fixture.GetUtcMonthWorks();
-                },
-                GetUtcDateWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "GetUtcDateWorks()", $t.Line = "316", $t));
-                    t.Fixture.GetUtcDateWorks();
-                },
-                GetUtcHoursWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "GetUtcHoursWorks()", $t.Line = "323", $t));
-                    t.Fixture.GetUtcHoursWorks();
-                },
-                GetUtcMinutesWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "GetUtcMinutesWorks()", $t.Line = "330", $t));
-                    t.Fixture.GetUtcMinutesWorks();
-                },
-                GetUtcSecondsWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "GetUtcSecondsWorks()", $t.Line = "337", $t));
-                    t.Fixture.GetUtcSecondsWorks();
-                },
-                GetUtcMillisecondsWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "GetUtcMillisecondsWorks()", $t.Line = "344", $t));
-                    t.Fixture.GetUtcMillisecondsWorks();
-                },
-                GetUtcDayWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "GetUtcDayWorks()", $t.Line = "351", $t));
-                    t.Fixture.GetUtcDayWorks();
-                },
-                ParseWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "ParseWorks()", $t.Line = "358", $t));
-                    t.Fixture.ParseWorks();
-                },
-                ParseExactWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "ParseExactWorks()", $t.Line = "367", $t));
-                    t.Fixture.ParseExactWorks();
-                },
-                ParseExactReturnsNullIfTheInputIsInvalid: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "ParseExactReturnsNullIfTheInputIsInvalid()", $t.Line = "378", $t));
-                    t.Fixture.ParseExactReturnsNullIfTheInputIsInvalid();
-                },
-                ParseExactWithCultureWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "ParseExactWithCultureWorks()", $t.Line = "384", $t));
-                    t.Fixture.ParseExactWithCultureWorks();
-                },
-                ParseExactWithCultureReturnsNullIfTheInputIsInvalid: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "ParseExactWithCultureReturnsNullIfTheInputIsInvalid()", $t.Line = "395", $t));
-                    t.Fixture.ParseExactWithCultureReturnsNullIfTheInputIsInvalid();
-                },
-                ParseExactUtcWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "ParseExactUtcWorks()", $t.Line = "401", $t));
-                    t.Fixture.ParseExactUtcWorks();
-                },
-                ParseExactUtcReturnsNullIfTheInputIsInvalid: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "ParseExactUtcReturnsNullIfTheInputIsInvalid()", $t.Line = "412", $t));
-                    t.Fixture.ParseExactUtcReturnsNullIfTheInputIsInvalid();
-                },
-                ParseExactUtcWithCultureWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "ParseExactUtcWithCultureWorks()", $t.Line = "418", $t));
-                    t.Fixture.ParseExactUtcWithCultureWorks();
-                },
-                ParseExactUtcWithCultureReturnsNullIfTheInputIsInvalid: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "ParseExactUtcWithCultureReturnsNullIfTheInputIsInvalid()", $t.Line = "429", $t));
-                    t.Fixture.ParseExactUtcWithCultureReturnsNullIfTheInputIsInvalid();
-                },
-                ToDateStringWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "ToDateStringWorks()", $t.Line = "435", $t));
-                    t.Fixture.ToDateStringWorks();
-                },
-                ToTimeStringWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "ToTimeStringWorks()", $t.Line = "443", $t));
-                    t.Fixture.ToTimeStringWorks();
-                },
-                ToUtcStringWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "ToUtcStringWorks()", $t.Line = "451", $t));
-                    t.Fixture.ToUtcStringWorks();
-                },
-                ToLocaleDateStringWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "ToLocaleDateStringWorks()", $t.Line = "459", $t));
-                    t.Fixture.ToLocaleDateStringWorks();
-                },
-                ToLocaleTimeStringWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "ToLocaleTimeStringWorks()", $t.Line = "467", $t));
-                    t.Fixture.ToLocaleTimeStringWorks();
-                },
-                UtcYMDWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "UtcYMDWorks()", $t.Line = "521", $t));
-                    t.Fixture.UtcYMDWorks();
-                },
-                UtcYMDHWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "UtcYMDHWorks()", $t.Line = "527", $t));
-                    t.Fixture.UtcYMDHWorks();
-                },
-                UtcYMDHNWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "UtcYMDHNWorks()", $t.Line = "533", $t));
-                    t.Fixture.UtcYMDHNWorks();
-                },
-                UtcYMDHNSWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "UtcYMDHNSWorks()", $t.Line = "539", $t));
-                    t.Fixture.UtcYMDHNSWorks();
-                },
-                UtcYMDHNSUWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "UtcYMDHNSUWorks()", $t.Line = "545", $t));
-                    t.Fixture.UtcYMDHNSUWorks();
-                },
-                SubtractingDatesWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "SubtractingDatesWorks()", $t.Line = "551", $t));
-                    t.Fixture.SubtractingDatesWorks();
-                },
-                SubtractMethodReturningTimeSpanWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "SubtractMethodReturningTimeSpanWorks()", $t.Line = "557", $t));
-                    t.Fixture.SubtractMethodReturningTimeSpanWorks();
-                },
-                DateEqualityWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "DateEqualityWorks()", $t.Line = "586", $t));
-                    t.Fixture.DateEqualityWorks();
-                },
-                DateInequalityWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "DateInequalityWorks()", $t.Line = "596", $t));
-                    t.Fixture.DateInequalityWorks();
-                },
-                DateLessThanWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "DateLessThanWorks()", $t.Line = "606", $t));
-                    t.Fixture.DateLessThanWorks();
-                },
-                DateLessEqualWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "DateLessEqualWorks()", $t.Line = "614", $t));
-                    t.Fixture.DateLessEqualWorks();
-                },
-                DateGreaterThanWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "DateGreaterThanWorks()", $t.Line = "622", $t));
-                    t.Fixture.DateGreaterThanWorks();
-                },
-                DateGreaterEqualWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "DateGreaterEqualWorks()", $t.Line = "630", $t));
-                    t.Fixture.DateGreaterEqualWorks();
-                },
-                ConvertingDateToMutableDateReturnsANewButEqualInstance: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "ConvertingDateToMutableDateReturnsANewButEqualInstance()", $t.Line = "638", $t));
-                    t.Fixture.ConvertingDateToMutableDateReturnsANewButEqualInstance();
-                },
-                ConvertingMutableDateToDateReturnsANewButEqualInstance: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "ConvertingMutableDateToDateReturnsANewButEqualInstance()", $t.Line = "649", $t));
-                    t.Fixture.ConvertingMutableDateToDateReturnsANewButEqualInstance();
-                },
-                GetHashCodeWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "GetHashCodeWorks()", $t.Line = "660", $t));
-                    t.Fixture.GetHashCodeWorks();
-                },
-                EqualsWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "EqualsWorks()", $t.Line = "669", $t));
-                    t.Fixture.EqualsWorks();
-                },
-                IEquatableEqualsWorks_SPI_1608: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "IEquatableEqualsWorks_SPI_1608()", $t.Line = "678", $t));
-                    t.Fixture.IEquatableEqualsWorks_SPI_1608();
-                },
-                StaticEqualsWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "StaticEqualsWorks()", $t.Line = "692", $t));
-                    t.Fixture.StaticEqualsWorks();
-                },
-                CompareToWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "CompareToWorks()", $t.Line = "701", $t));
-                    t.Fixture.CompareToWorks();
-                },
-                StaticCompareWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "StaticCompareWorks()", $t.Line = "709", $t));
-                    t.Fixture.StaticCompareWorks();
-                },
-                IComparableCompareToWorks_SPI_1609: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "IComparableCompareToWorks_SPI_1609()", $t.Line = "717", $t));
-                    t.Fixture.IComparableCompareToWorks_SPI_1609();
-                },
-                DatePropertyWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "DatePropertyWorks()", $t.Line = "726", $t));
-                    t.Fixture.DatePropertyWorks();
-                },
-                DayPropertyWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "DayPropertyWorks()", $t.Line = "733", $t));
-                    t.Fixture.DayPropertyWorks();
-                },
-                DayOfWeekPropertyWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "DayOfWeekPropertyWorks()", $t.Line = "740", $t));
-                    t.Fixture.DayOfWeekPropertyWorks();
-                },
-                DayOfYearPropertyWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "DayOfYearPropertyWorks()", $t.Line = "747", $t));
-                    t.Fixture.DayOfYearPropertyWorks();
-                },
-                HourPropertyWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "HourPropertyWorks()", $t.Line = "754", $t));
-                    t.Fixture.HourPropertyWorks();
-                },
-                MillisecondPropertyWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "MillisecondPropertyWorks()", $t.Line = "761", $t));
-                    t.Fixture.MillisecondPropertyWorks();
-                },
-                MinutePropertyWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "MinutePropertyWorks()", $t.Line = "768", $t));
-                    t.Fixture.MinutePropertyWorks();
-                },
-                MonthPropertyWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "MonthPropertyWorks()", $t.Line = "775", $t));
-                    t.Fixture.MonthPropertyWorks();
-                },
-                SecondPropertyWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "SecondPropertyWorks()", $t.Line = "782", $t));
-                    t.Fixture.SecondPropertyWorks();
-                },
-                YearPropertyWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "YearPropertyWorks()", $t.Line = "789", $t));
-                    t.Fixture.YearPropertyWorks();
-                },
-                AddDaysWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "AddDaysWorks()", $t.Line = "796", $t));
-                    t.Fixture.AddDaysWorks();
-                },
-                AddHoursWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "AddHoursWorks()", $t.Line = "805", $t));
-                    t.Fixture.AddHoursWorks();
-                },
-                AddMillisecondsWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "AddMillisecondsWorks()", $t.Line = "814", $t));
-                    t.Fixture.AddMillisecondsWorks();
-                },
-                AddMinutesWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "AddMinutesWorks()", $t.Line = "823", $t));
-                    t.Fixture.AddMinutesWorks();
-                },
-                AddMonthsWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "AddMonthsWorks()", $t.Line = "832", $t));
-                    t.Fixture.AddMonthsWorks();
-                },
-                AddSecondsWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "AddSecondsWorks()", $t.Line = "841", $t));
-                    t.Fixture.AddSecondsWorks();
-                },
-                AddYearsWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "AddYearsWorks()", $t.Line = "850", $t));
-                    t.Fixture.AddYearsWorks();
-                },
-                DaysInMonthWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "DaysInMonthWorks()", $t.Line = "859", $t));
-                    t.Fixture.DaysInMonthWorks();
-                },
-                IsLeapYearWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DateTimeTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "IsLeapYearWorks()", $t.Line = "878", $t));
-                    t.Fixture.IsLeapYearWorks();
-                }
-            }
-        },
-        fields: {
-            context: null
-        },
-        methods: {
-            GetContext: function () {
-                var $t;
-                if (this.context == null) {
-                    this.context = ($t = new Bridge.Test.Runtime.FixtureContext(), $t.Project = "Batch4", $t.ClassName = "Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests", $t.File = "Batch4\\SimpleTypes\\DateTimeTests.cs", $t);
                 }
                 return this.context;
             }
