+++ Tests/Runner/Batch4/code.js
@@ -17042,549 +17042,6 @@ Bridge.assembly("Bridge.ClientTest.Batch4", {"Bridge.ClientTest.Batch4.Reflectio
         }
     });
 
-    Bridge.define("Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests", {
-        methods: {
-            TypePropertiesAreCorrect_SPI_1607_1608_1609: function () {
-                Bridge.Test.NUnit.Assert.AreEqual("System.DateTime", Bridge.Reflection.getTypeFullName(System.DateTime));
-                Bridge.Test.NUnit.Assert.False(Bridge.Reflection.isClass(System.DateTime));
-                // #1607 #1608 #1609
-                Bridge.Test.NUnit.Assert.True(Bridge.Reflection.isAssignableFrom(System.IComparable$1(System.DateTime), System.DateTime));
-                Bridge.Test.NUnit.Assert.True(Bridge.Reflection.isAssignableFrom(System.IEquatable$1(System.DateTime), System.DateTime));
-                Bridge.Test.NUnit.Assert.True(Bridge.Reflection.isAssignableFrom(System.IFormattable, System.DateTime));
-
-                var d = Bridge.box(System.DateTime.getDefaultValue(), System.DateTime, System.DateTime.format);
-                Bridge.Test.NUnit.Assert.True(Bridge.is(d, System.DateTime));
-                // #1609
-                Bridge.Test.NUnit.Assert.True(Bridge.is(d, System.IComparable$1(System.DateTime)));
-                // #1608
-                Bridge.Test.NUnit.Assert.True(Bridge.is(d, System.IEquatable$1(System.DateTime)));
-
-                Bridge.Test.NUnit.Assert.True(Bridge.is(d, System.IFormattable));
-
-                var interfaces = Bridge.Reflection.getInterfaces(System.DateTime);
-                Bridge.Test.NUnit.Assert.AreEqual(4, interfaces.length);
-                Bridge.Test.NUnit.Assert.True(System.Array.contains(interfaces, System.IComparable$1(System.DateTime), Function));
-                Bridge.Test.NUnit.Assert.True(System.Array.contains(interfaces, System.IEquatable$1(System.DateTime), Function));
-                Bridge.Test.NUnit.Assert.True(System.Array.contains(interfaces, System.IFormattable, Function));
-            },
-            DefaultConstructorWorks_SPI_1606: function () {
-                var dt = System.DateTime.getDefaultValue();
-                // #1606
-                Bridge.Test.NUnit.Assert.AreEqual(1, dt.getFullYear());
-            },
-            DefaultValueWorks_SPI_1606: function () {
-                var dt = Bridge.getDefaultValue(System.DateTime);
-                // #1606
-                Bridge.Test.NUnit.Assert.AreEqual(1, dt.getFullYear());
-            },
-            CreatingInstanceReturnsDateWithZeroValue_SPI_1606: function () {
-                var dt = Bridge.createInstance(System.DateTime);
-                // #1606
-                Bridge.Test.NUnit.Assert.AreEqual(1, dt.getFullYear());
-            },
-            MillisecondSinceEpochConstructorWorks: function () {
-                var dt = System.DateTime.fromTicks(System.Int64([250327040,10]));
-                Bridge.Test.NUnit.Assert.AreEqual(1, new Date(dt.valueOf() + Math.round((1) * 864e5)).getFullYear());
-            },
-            StringConstructorWorks: function () {
-                var dt = new Date("Aug 12, 2012");
-                Bridge.Test.NUnit.Assert.AreEqual(2012, dt.getFullYear());
-                Bridge.Test.NUnit.Assert.AreEqual(8, (dt.getMonth() + 1));
-                Bridge.Test.NUnit.Assert.AreEqual(12, dt.getDate());
-            },
-            YMDConstructorWorks: function () {
-                var dt = new Date(2011, 7 - 1, 12);
-                Bridge.Test.NUnit.Assert.AreEqual(2011, dt.getFullYear());
-                Bridge.Test.NUnit.Assert.AreEqual(7, (dt.getMonth() + 1));
-                Bridge.Test.NUnit.Assert.AreEqual(12, dt.getDate());
-            },
-            YMDHConstructorWorks: function () {
-                var dt = new Date(2011, 7 - 1, 12, 13);
-                Bridge.Test.NUnit.Assert.AreEqual(2011, dt.getFullYear());
-                Bridge.Test.NUnit.Assert.AreEqual(7, (dt.getMonth() + 1));
-                Bridge.Test.NUnit.Assert.AreEqual(12, dt.getDate());
-                Bridge.Test.NUnit.Assert.AreEqual(13, dt.getHours());
-            },
-            YMDHNConstructorWorks: function () {
-                var dt = new Date(2011, 7 - 1, 12, 13, 42);
-                Bridge.Test.NUnit.Assert.AreEqual(2011, dt.getFullYear());
-                Bridge.Test.NUnit.Assert.AreEqual(7, (dt.getMonth() + 1));
-                Bridge.Test.NUnit.Assert.AreEqual(12, dt.getDate());
-                Bridge.Test.NUnit.Assert.AreEqual(13, dt.getHours());
-                Bridge.Test.NUnit.Assert.AreEqual(42, dt.getMinutes());
-            },
-            YMDHNSConstructorWorks: function () {
-                var dt = new Date(2011, 7 - 1, 12, 13, 42, 56);
-                Bridge.Test.NUnit.Assert.AreEqual(2011, dt.getFullYear());
-                Bridge.Test.NUnit.Assert.AreEqual(7, (dt.getMonth() + 1));
-                Bridge.Test.NUnit.Assert.AreEqual(12, dt.getDate());
-                Bridge.Test.NUnit.Assert.AreEqual(13, dt.getHours());
-                Bridge.Test.NUnit.Assert.AreEqual(42, dt.getMinutes());
-                Bridge.Test.NUnit.Assert.AreEqual(56, dt.getSeconds());
-            },
-            YMDHNSUConstructorWorks: function () {
-                var dt = new Date(2011, 7 - 1, 12, 13, 42, 56, 345);
-                Bridge.Test.NUnit.Assert.AreEqual(2011, dt.getFullYear());
-                Bridge.Test.NUnit.Assert.AreEqual(7, (dt.getMonth() + 1));
-                Bridge.Test.NUnit.Assert.AreEqual(12, dt.getDate());
-                Bridge.Test.NUnit.Assert.AreEqual(13, dt.getHours());
-                Bridge.Test.NUnit.Assert.AreEqual(42, dt.getMinutes());
-                Bridge.Test.NUnit.Assert.AreEqual(56, dt.getSeconds());
-                Bridge.Test.NUnit.Assert.AreEqual(345, dt.getMilliseconds());
-            },
-            NowWorks: function () {
-                var dt = new Date();
-                Bridge.Test.NUnit.Assert.True(dt.getFullYear() > 2011);
-            },
-            UtcNowWorks: function () {
-                var utc = System.DateTime.utcNow();
-                var local = new Date();
-                var date = new Date(local.getUTCFullYear(), (local.getUTCMonth() + 1) - 1, local.getUTCDate(), local.getUTCHours(), local.getUTCMinutes(), local.getUTCSeconds(), local.getUTCMilliseconds());
-                var diff = System.DateTime.subdd(date, utc);
-                Bridge.Test.NUnit.Assert.True(Math.abs(diff.getSeconds()) < 1000);
-            },
-            ToUniversalWorks: function () {
-                var dt = new Date(2011, 7 - 1, 12, 13, 42, 56, 345);
-                var utc = System.DateTime.toUTC(dt);
-                Bridge.Test.NUnit.Assert.AreEqual(utc.getFullYear(), dt.getUTCFullYear());
-                Bridge.Test.NUnit.Assert.AreEqual((utc.getMonth() + 1), (dt.getUTCMonth() + 1));
-                Bridge.Test.NUnit.Assert.AreEqual(utc.getDate(), dt.getUTCDate());
-                Bridge.Test.NUnit.Assert.AreEqual(utc.getHours(), dt.getUTCHours());
-                Bridge.Test.NUnit.Assert.AreEqual(utc.getMinutes(), dt.getUTCMinutes());
-                Bridge.Test.NUnit.Assert.AreEqual(utc.getSeconds(), dt.getUTCSeconds());
-                Bridge.Test.NUnit.Assert.AreEqual(utc.getMilliseconds(), dt.getUTCMilliseconds());
-            },
-            ToLocalWorks: function () {
-                var utc = new Date(2011, 7 - 1, 12, 13, 42, 56, 345);
-                var dt = System.DateTime.toLocal(utc);
-                Bridge.Test.NUnit.Assert.AreEqual(utc.getFullYear(), dt.getUTCFullYear());
-                Bridge.Test.NUnit.Assert.AreEqual((utc.getMonth() + 1), (dt.getUTCMonth() + 1));
-                Bridge.Test.NUnit.Assert.AreEqual(utc.getDate(), dt.getUTCDate());
-                Bridge.Test.NUnit.Assert.AreEqual(utc.getHours(), dt.getUTCHours());
-                Bridge.Test.NUnit.Assert.AreEqual(utc.getMinutes(), dt.getUTCMinutes());
-                Bridge.Test.NUnit.Assert.AreEqual(utc.getSeconds(), dt.getUTCSeconds());
-                Bridge.Test.NUnit.Assert.AreEqual(utc.getMilliseconds(), dt.getUTCMilliseconds());
-            },
-            TodayWorks: function () {
-                var dt = System.DateTime.today();
-                Bridge.Test.NUnit.Assert.True(dt.getFullYear() > 2011);
-                Bridge.Test.NUnit.Assert.AreEqual(0, dt.getHours());
-                Bridge.Test.NUnit.Assert.AreEqual(0, dt.getMinutes());
-                Bridge.Test.NUnit.Assert.AreEqual(0, dt.getSeconds());
-                Bridge.Test.NUnit.Assert.AreEqual(0, dt.getMilliseconds());
-            },
-            FormatWorks: function () {
-                var dt = new Date(2011, 7 - 1, 12, 13);
-                Bridge.Test.NUnit.Assert.AreEqual("2011-07-12", System.DateTime.format(dt, "yyyy-MM-dd"));
-            },
-            ToStringWithFormatWorks: function () {
-                var dt = new Date(2011, 7 - 1, 12);
-                Bridge.Test.NUnit.Assert.AreEqual("2011-07-12", System.DateTime.format(dt, "yyyy-MM-dd"));
-            },
-            ToStringWithFormatAndProviderWorks: function () {
-                var dt = new Date(2011, 7 - 1, 12);
-                Bridge.Test.NUnit.Assert.AreEqual("2011-07-12", System.DateTime.format(dt, "yyyy-MM-dd", System.Globalization.CultureInfo.invariantCulture));
-            },
-            IFormattableToStringWorks: function () {
-                var dt = new Date(2011, 7 - 1, 12);
-                Bridge.Test.NUnit.Assert.AreEqual("2011-07-12", Bridge.format(Bridge.cast(Bridge.box(dt, System.DateTime, System.DateTime.format), System.IFormattable), "yyyy-MM-dd", System.Globalization.CultureInfo.invariantCulture));
-            },
-            GetFullYearWorks: function () {
-                var dt = new Date(2011, 7 - 1, 12, 13, 42, 56, 345);
-                Bridge.Test.NUnit.Assert.AreEqual(2011, dt.getFullYear());
-            },
-            GetMonthWorks: function () {
-                var dt = new Date(2011, 7 - 1, 12, 13, 42, 56, 345);
-                Bridge.Test.NUnit.Assert.AreEqual(7, (dt.getMonth() + 1));
-            },
-            GetDateWorks: function () {
-                var dt = new Date(2011, 7 - 1, 12, 13, 42, 56, 345);
-                Bridge.Test.NUnit.Assert.AreEqual(12, dt.getDate());
-            },
-            GetHoursWorks: function () {
-                var dt = new Date(2011, 7 - 1, 12, 13, 42, 56, 345);
-                Bridge.Test.NUnit.Assert.AreEqual(13, dt.getHours());
-            },
-            GetMinutesWorks: function () {
-                var dt = new Date(2011, 7 - 1, 12, 13, 42, 56, 345);
-                Bridge.Test.NUnit.Assert.AreEqual(42, dt.getMinutes());
-            },
-            GetSecondsWorks: function () {
-                var dt = new Date(2011, 7 - 1, 12, 13, 42, 56, 345);
-                Bridge.Test.NUnit.Assert.AreEqual(56, dt.getSeconds());
-            },
-            GetMillisecondsWorks: function () {
-                var dt = new Date(2011, 7 - 1, 12, 13, 42, 56, 345);
-                Bridge.Test.NUnit.Assert.AreEqual(345, dt.getMilliseconds());
-            },
-            GetDayWorks: function () {
-                var dt = new Date(2011, 8 - 1, 12, 13, 42, 56, 345);
-                Bridge.Test.NUnit.Assert.AreEqual(5, dt.getDay());
-            },
-            GetTimeWorks: function () {
-                var dt = System.DateTime.fromTicks(System.DateTime.utc(1000, 1, 2));
-                Bridge.Test.NUnit.Assert.AreEqual((System.Int64([94318592,-7127])).toString(), dt.getTime().toString());
-            },
-            ValueOfWorks: function () {
-                var dt = System.DateTime.fromTicks(System.DateTime.utc(1000, 1, 2));
-                Bridge.Test.NUnit.Assert.AreEqual((System.Int64([94318592,-7127])).toString(), dt.valueOf().toString());
-            },
-            GetUtcFullYearWorks: function () {
-                var dt = System.DateTime.fromTicks(System.DateTime.utc(2011, 7, 12, 13, 42, 56, 345));
-                Bridge.Test.NUnit.Assert.AreEqual(2011, dt.getUTCFullYear());
-            },
-            GetUtcMonthWorks: function () {
-                var dt = System.DateTime.fromTicks(System.DateTime.utc(2011, 7, 12, 13, 42, 56, 345));
-                Bridge.Test.NUnit.Assert.AreEqual(7, (dt.getUTCMonth() + 1));
-            },
-            GetUtcDateWorks: function () {
-                var dt = System.DateTime.fromTicks(System.DateTime.utc(2011, 7, 12, 13, 42, 56, 345));
-                Bridge.Test.NUnit.Assert.AreEqual(12, dt.getUTCDate());
-            },
-            GetUtcHoursWorks: function () {
-                var dt = System.DateTime.fromTicks(System.DateTime.utc(2011, 7, 12, 13, 42, 56, 345));
-                Bridge.Test.NUnit.Assert.AreEqual(13, dt.getUTCHours());
-            },
-            GetUtcMinutesWorks: function () {
-                var dt = System.DateTime.fromTicks(System.DateTime.utc(2011, 7, 12, 13, 42, 56, 345));
-                Bridge.Test.NUnit.Assert.AreEqual(42, dt.getUTCMinutes());
-            },
-            GetUtcSecondsWorks: function () {
-                var dt = System.DateTime.fromTicks(System.DateTime.utc(2011, 7, 12, 13, 42, 56, 345));
-                Bridge.Test.NUnit.Assert.AreEqual(56, dt.getUTCSeconds());
-            },
-            GetUtcMillisecondsWorks: function () {
-                var dt = System.DateTime.fromTicks(System.DateTime.utc(2011, 7, 12, 13, 42, 56, 345));
-                Bridge.Test.NUnit.Assert.AreEqual(345, dt.getUTCMilliseconds());
-            },
-            GetUtcDayWorks: function () {
-                var dt = System.DateTime.fromTicks(System.DateTime.utc(2011, 8, 12, 13, 42, 56, 345));
-                Bridge.Test.NUnit.Assert.AreEqual(5, dt.getUTCDay());
-            },
-            ParseWorks: function () {
-                var dt = System.DateTime.parse("Aug 12, 2012");
-                Bridge.Test.NUnit.Assert.AreEqual(2012, dt.getFullYear());
-                Bridge.Test.NUnit.Assert.AreEqual(8, (dt.getMonth() + 1));
-                Bridge.Test.NUnit.Assert.AreEqual(12, dt.getDate());
-            },
-            ParseExactWorks: function () {
-                var dt = System.DateTime.parseExact("2012-12-08", "yyyy-dd-MM");
-                // Not C# API
-                //Assert.True(dt.HasValue);
-                Bridge.Test.NUnit.Assert.AreEqual(2012, dt.getFullYear());
-                Bridge.Test.NUnit.Assert.AreEqual(8, (dt.getMonth() + 1));
-                Bridge.Test.NUnit.Assert.AreEqual(12, dt.getDate());
-            },
-            ParseExactReturnsNullIfTheInputIsInvalid: function () {
-                Bridge.Test.NUnit.Assert.Throws$2(System.FormatException, $asm.$.Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests.f1);
-            },
-            ParseExactWithCultureWorks: function () {
-                var dt = System.DateTime.parseExact("2012-12-08", "yyyy-dd-MM", System.Globalization.CultureInfo.invariantCulture);
-                // Not C# API
-                //Assert.True(dt.HasValue);
-                Bridge.Test.NUnit.Assert.AreEqual(2012, dt.getFullYear());
-                Bridge.Test.NUnit.Assert.AreEqual(8, (dt.getMonth() + 1));
-                Bridge.Test.NUnit.Assert.AreEqual(12, dt.getDate());
-            },
-            ParseExactWithCultureReturnsNullIfTheInputIsInvalid: function () {
-                Bridge.Test.NUnit.Assert.Throws$2(System.FormatException, $asm.$.Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests.f2);
-            },
-            ParseExactUtcWorks: function () {
-                var dt = System.DateTime.parseExact("2012-12-08", "yyyy-dd-MM", null, true);
-                // Not C# API
-                //Assert.True(dt.HasValue);
-                Bridge.Test.NUnit.Assert.AreEqual(2012, dt.getUTCFullYear());
-                Bridge.Test.NUnit.Assert.AreEqual(8, (dt.getUTCMonth() + 1));
-                Bridge.Test.NUnit.Assert.AreEqual(12, dt.getUTCDate());
-            },
-            ParseExactUtcReturnsNullIfTheInputIsInvalid: function () {
-                Bridge.Test.NUnit.Assert.Throws$2(System.FormatException, $asm.$.Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests.f3);
-            },
-            ParseExactUtcWithCultureWorks: function () {
-                var dt = System.DateTime.parseExact("2012-12-08", "yyyy-dd-MM", System.Globalization.CultureInfo.invariantCulture, true);
-                // Not C# API
-                //Assert.True(dt.HasValue);
-                Bridge.Test.NUnit.Assert.AreEqual(2012, dt.getUTCFullYear());
-                Bridge.Test.NUnit.Assert.AreEqual(8, (dt.getUTCMonth() + 1));
-                Bridge.Test.NUnit.Assert.AreEqual(12, dt.getUTCDate());
-            },
-            ParseExactUtcWithCultureReturnsNullIfTheInputIsInvalid: function () {
-                Bridge.Test.NUnit.Assert.Throws$2(System.FormatException, $asm.$.Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests.f4);
-            },
-            ToDateStringWorks: function () {
-                var dt = new Date(2011, 7 - 1, 12, 13, 42);
-                var s = dt.toDateString();
-                Bridge.Test.NUnit.Assert.True(System.String.indexOf(s, "2011") >= 0 && System.String.indexOf(s, "42") < 0);
-            },
-            ToTimeStringWorks: function () {
-                var dt = new Date(2011, 7 - 1, 12, 13, 42);
-                var s = dt.toTimeString();
-                Bridge.Test.NUnit.Assert.True(System.String.indexOf(s, "2011") < 0 && System.String.indexOf(s, "42") >= 0);
-            },
-            ToUtcStringWorks: function () {
-                var dt = new Date(2011, 7 - 1, 12, 13, 42);
-                var s = dt.toUTCString();
-                Bridge.Test.NUnit.Assert.True(System.String.indexOf(s, "2011") >= 0 && System.String.indexOf(s, "42") >= 0);
-            },
-            ToLocaleDateStringWorks: function () {
-                var dt = new Date(2011, 7 - 1, 12, 13, 42);
-                var s = dt.toLocaleDateString();
-                Bridge.Test.NUnit.Assert.True(System.String.indexOf(s, "2011") >= 0 && System.String.indexOf(s, "42") < 0);
-            },
-            ToLocaleTimeStringWorks: function () {
-                var dt = new Date(2011, 7 - 1, 12, 13, 42);
-                var s = dt.toLocaleTimeString();
-                Bridge.Test.NUnit.Assert.True(System.String.indexOf(s, "2011") < 0 && System.String.indexOf(s, "42") >= 0);
-            },
-            AssertDateUtc: function (dt, year, month, day, hours, minutes, seconds, milliseconds) {
-                Bridge.Test.NUnit.Assert.AreEqual(year, dt.getUTCFullYear());
-                Bridge.Test.NUnit.Assert.AreEqual(month, (dt.getUTCMonth() + 1));
-                Bridge.Test.NUnit.Assert.AreEqual(day, dt.getUTCDate());
-                Bridge.Test.NUnit.Assert.AreEqual(hours, dt.getUTCHours());
-                Bridge.Test.NUnit.Assert.AreEqual(minutes, dt.getUTCMinutes());
-                Bridge.Test.NUnit.Assert.AreEqual(seconds, dt.getUTCSeconds());
-                Bridge.Test.NUnit.Assert.AreEqual(milliseconds, dt.getUTCMilliseconds());
-            },
-            UtcYMDWorks: function () {
-                this.AssertDateUtc(System.DateTime.fromTicks(System.DateTime.utc(2011, 7, 12)), 2011, 7, 12, 0, 0, 0, 0);
-            },
-            UtcYMDHWorks: function () {
-                this.AssertDateUtc(System.DateTime.fromTicks(System.DateTime.utc(2011, 7, 12, 13)), 2011, 7, 12, 13, 0, 0, 0);
-            },
-            UtcYMDHNWorks: function () {
-                this.AssertDateUtc(System.DateTime.fromTicks(System.DateTime.utc(2011, 7, 12, 13, 42)), 2011, 7, 12, 13, 42, 0, 0);
-            },
-            UtcYMDHNSWorks: function () {
-                this.AssertDateUtc(System.DateTime.fromTicks(System.DateTime.utc(2011, 7, 12, 13, 42, 56)), 2011, 7, 12, 13, 42, 56, 0);
-            },
-            UtcYMDHNSUWorks: function () {
-                this.AssertDateUtc(System.DateTime.fromTicks(System.DateTime.utc(2011, 7, 12, 13, 42, 56, 345)), 2011, 7, 12, 13, 42, 56, 345);
-            },
-            SubtractingDatesWorks: function () {
-                Bridge.Test.NUnit.Assert.AreEqual(86400000, (System.DateTime.subdd(new Date(2011, 7 - 1, 12), new Date(2011, 7 - 1, 11))).getTotalMilliseconds());
-            },
-            SubtractMethodReturningTimeSpanWorks: function () {
-                Bridge.Test.NUnit.Assert.AreEqual(new System.TimeSpan(1, 0, 0, 0), System.DateTime.subdd(new Date(2011, 7 - 1, 12), new Date(2011, 7 - 1, 11)));
-                Bridge.Test.NUnit.Assert.AreEqual(new System.TimeSpan(1, 2, 0, 0), System.DateTime.subdd(new Date(2011, 7 - 1, 12, 15, 0, 0), new Date(2011, 7 - 1, 11, 13, 0, 0)));
-            },
-            DateEqualityWorks: function () {
-                Bridge.Test.NUnit.Assert.True(Bridge.equals(new Date(2011, 7 - 1, 12), new Date(2011, 7 - 1, 12)));
-                Bridge.Test.NUnit.Assert.False(Bridge.equals(new Date(2011, 7 - 1, 12), new Date(2011, 7 - 1, 13)));
-                Bridge.Test.NUnit.Assert.AreStrictEqual(false, Bridge.equals(new Date(2011, 7 - 1, 12), null));
-                Bridge.Test.NUnit.Assert.AreStrictEqual(false, Bridge.equals(null, new Date(2011, 7 - 1, 12)));
-                Bridge.Test.NUnit.Assert.AreStrictEqual(true, Bridge.equals(null, null));
-            },
-            DateInequalityWorks: function () {
-                Bridge.Test.NUnit.Assert.False(!Bridge.equals(new Date(2011, 7 - 1, 12), new Date(2011, 7 - 1, 12)));
-                Bridge.Test.NUnit.Assert.True(!Bridge.equals(new Date(2011, 7 - 1, 12), new Date(2011, 7 - 1, 13)));
-                Bridge.Test.NUnit.Assert.AreStrictEqual(true, !Bridge.equals(new Date(2011, 7 - 1, 12), null));
-                Bridge.Test.NUnit.Assert.AreStrictEqual(true, !Bridge.equals(null, new Date(2011, 7 - 1, 12)));
-                Bridge.Test.NUnit.Assert.AreStrictEqual(false, !Bridge.equals(null, null));
-            },
-            DateLessThanWorks: function () {
-                Bridge.Test.NUnit.Assert.True(System.DateTime.lt(new Date(2011, 7 - 1, 11), new Date(2011, 7 - 1, 12)));
-                Bridge.Test.NUnit.Assert.False(System.DateTime.lt(new Date(2011, 7 - 1, 12), new Date(2011, 7 - 1, 12)));
-                Bridge.Test.NUnit.Assert.False(System.DateTime.lt(new Date(2011, 7 - 1, 13), new Date(2011, 7 - 1, 12)));
-            },
-            DateLessEqualWorks: function () {
-                Bridge.Test.NUnit.Assert.True(System.DateTime.lte(new Date(2011, 7 - 1, 11), new Date(2011, 7 - 1, 12)));
-                Bridge.Test.NUnit.Assert.True(System.DateTime.lte(new Date(2011, 7 - 1, 12), new Date(2011, 7 - 1, 12)));
-                Bridge.Test.NUnit.Assert.False(System.DateTime.lte(new Date(2011, 7 - 1, 13), new Date(2011, 7 - 1, 12)));
-            },
-            DateGreaterThanWorks: function () {
-                Bridge.Test.NUnit.Assert.False(System.DateTime.gt(new Date(2011, 7 - 1, 11), new Date(2011, 7 - 1, 12)));
-                Bridge.Test.NUnit.Assert.False(System.DateTime.gt(new Date(2011, 7 - 1, 12), new Date(2011, 7 - 1, 12)));
-                Bridge.Test.NUnit.Assert.True(System.DateTime.gt(new Date(2011, 7 - 1, 13), new Date(2011, 7 - 1, 12)));
-            },
-            DateGreaterEqualWorks: function () {
-                Bridge.Test.NUnit.Assert.False(System.DateTime.gte(new Date(2011, 7 - 1, 11), new Date(2011, 7 - 1, 12)));
-                Bridge.Test.NUnit.Assert.True(System.DateTime.gte(new Date(2011, 7 - 1, 12), new Date(2011, 7 - 1, 12)));
-                Bridge.Test.NUnit.Assert.True(System.DateTime.gte(new Date(2011, 7 - 1, 13), new Date(2011, 7 - 1, 12)));
-            },
-            ConvertingDateToMutableDateReturnsANewButEqualInstance: function () {
-                var dt = new Date(2011, 7 - 1, 12);
-                var mdt = new Date(dt.valueOf());
-                Bridge.Test.NUnit.Assert.False(Bridge.referenceEquals(Bridge.box(dt, System.DateTime, System.DateTime.format), mdt));
-                Bridge.Test.NUnit.Assert.AreEqual(2011, mdt.getFullYear());
-                Bridge.Test.NUnit.Assert.AreEqual(6, mdt.getMonth());
-                Bridge.Test.NUnit.Assert.AreEqual(12, mdt.getDate());
-            },
-            ConvertingMutableDateToDateReturnsANewButEqualInstance: function () {
-                var mdt = new Date(2011, 7, 12);
-                var dt = new Date(mdt.valueOf());
-                Bridge.Test.NUnit.Assert.False(Bridge.referenceEquals(Bridge.box(dt, System.DateTime, System.DateTime.format), mdt));
-                Bridge.Test.NUnit.Assert.AreEqual(2011, mdt.getFullYear());
-                Bridge.Test.NUnit.Assert.AreEqual(7, mdt.getMonth());
-                Bridge.Test.NUnit.Assert.AreEqual(12, mdt.getDate());
-            },
-            GetHashCodeWorks: function () {
-                Bridge.Test.NUnit.Assert.AreEqual(Bridge.getHashCode(System.DateTime.fromTicks(System.Int64(0))), Bridge.getHashCode(System.DateTime.fromTicks(System.Int64(0))));
-                Bridge.Test.NUnit.Assert.AreEqual(Bridge.getHashCode(System.DateTime.fromTicks(System.Int64(10000))), Bridge.getHashCode(System.DateTime.fromTicks(System.Int64(10000))));
-                Bridge.Test.NUnit.Assert.AreNotEqual(Bridge.getHashCode(System.DateTime.fromTicks(System.Int64(10000))), Bridge.getHashCode(System.DateTime.fromTicks(System.Int64(0))));
-                Bridge.Test.NUnit.Assert.True(System.Int64(Bridge.getHashCode(new Date(3000, 1 - 1, 1))).lt(System.Int64([-1,0])));
-            },
-            EqualsWorks: function () {
-                Bridge.Test.NUnit.Assert.True(Bridge.equals(System.DateTime.fromTicks(System.Int64(0)), Bridge.box(System.DateTime.fromTicks(System.Int64(0)), System.DateTime, System.DateTime.format)));
-                Bridge.Test.NUnit.Assert.False(Bridge.equals(System.DateTime.fromTicks(System.Int64(10000)), Bridge.box(System.DateTime.fromTicks(System.Int64(0)), System.DateTime, System.DateTime.format)));
-                Bridge.Test.NUnit.Assert.False(Bridge.equals(System.DateTime.fromTicks(System.Int64(0)), Bridge.box(System.DateTime.fromTicks(System.Int64(10000)), System.DateTime, System.DateTime.format)));
-                Bridge.Test.NUnit.Assert.True(Bridge.equals(System.DateTime.fromTicks(System.Int64(10000)), Bridge.box(System.DateTime.fromTicks(System.Int64(10000)), System.DateTime, System.DateTime.format)));
-            },
-            IEquatableEqualsWorks_SPI_1608: function () {
-                Bridge.Test.NUnit.Assert.True(Bridge.equalsT(System.DateTime.fromTicks(System.Int64(0)), System.DateTime.fromTicks(System.Int64(0))));
-                Bridge.Test.NUnit.Assert.False(Bridge.equalsT(System.DateTime.fromTicks(System.Int64(10000)), System.DateTime.fromTicks(System.Int64(0))));
-                Bridge.Test.NUnit.Assert.False(Bridge.equalsT(System.DateTime.fromTicks(System.Int64(0)), System.DateTime.fromTicks(System.Int64(10000))));
-                Bridge.Test.NUnit.Assert.True(Bridge.equalsT(System.DateTime.fromTicks(System.Int64(10000)), System.DateTime.fromTicks(System.Int64(10000))));
-
-                Bridge.Test.NUnit.Assert.True(Bridge.equalsT(Bridge.cast(Bridge.box(System.DateTime.fromTicks(System.Int64(0)), System.DateTime, System.DateTime.format), System.IEquatable$1(System.DateTime)), System.DateTime.fromTicks(System.Int64(0)), System.DateTime));
-                Bridge.Test.NUnit.Assert.False(Bridge.equalsT(Bridge.cast(Bridge.box(System.DateTime.fromTicks(System.Int64(10000)), System.DateTime, System.DateTime.format), System.IEquatable$1(System.DateTime)), System.DateTime.fromTicks(System.Int64(0)), System.DateTime));
-                Bridge.Test.NUnit.Assert.False(Bridge.equalsT(Bridge.cast(Bridge.box(System.DateTime.fromTicks(System.Int64(0)), System.DateTime, System.DateTime.format), System.IEquatable$1(System.DateTime)), System.DateTime.fromTicks(System.Int64(10000)), System.DateTime));
-                Bridge.Test.NUnit.Assert.True(Bridge.equalsT(Bridge.cast(Bridge.box(System.DateTime.fromTicks(System.Int64(10000)), System.DateTime, System.DateTime.format), System.IEquatable$1(System.DateTime)), System.DateTime.fromTicks(System.Int64(10000)), System.DateTime));
-            },
-            StaticEqualsWorks: function () {
-                Bridge.Test.NUnit.Assert.True(Bridge.equalsT(System.DateTime.fromTicks(System.Int64(0)), System.DateTime.fromTicks(System.Int64(0))));
-                Bridge.Test.NUnit.Assert.False(Bridge.equalsT(System.DateTime.fromTicks(System.Int64(10000)), System.DateTime.fromTicks(System.Int64(0))));
-                Bridge.Test.NUnit.Assert.False(Bridge.equalsT(System.DateTime.fromTicks(System.Int64(0)), System.DateTime.fromTicks(System.Int64(10000))));
-                Bridge.Test.NUnit.Assert.True(Bridge.equalsT(System.DateTime.fromTicks(System.Int64(10000)), System.DateTime.fromTicks(System.Int64(10000))));
-            },
-            CompareToWorks: function () {
-                Bridge.Test.NUnit.Assert.True(Bridge.compare(System.DateTime.fromTicks(System.Int64(0)), System.DateTime.fromTicks(System.Int64(0))) === 0);
-                Bridge.Test.NUnit.Assert.True(Bridge.compare(System.DateTime.fromTicks(System.Int64(10000)), System.DateTime.fromTicks(System.Int64(0))) > 0);
-                Bridge.Test.NUnit.Assert.True(Bridge.compare(System.DateTime.fromTicks(System.Int64(0)), System.DateTime.fromTicks(System.Int64(10000))) < 0);
-            },
-            StaticCompareWorks: function () {
-                Bridge.Test.NUnit.Assert.True(Bridge.compare(System.DateTime.fromTicks(System.Int64(0)), System.DateTime.fromTicks(System.Int64(0))) === 0);
-                Bridge.Test.NUnit.Assert.True(Bridge.compare(System.DateTime.fromTicks(System.Int64(10000)), System.DateTime.fromTicks(System.Int64(0))) > 0);
-                Bridge.Test.NUnit.Assert.True(Bridge.compare(System.DateTime.fromTicks(System.Int64(0)), System.DateTime.fromTicks(System.Int64(10000))) < 0);
-            },
-            IComparableCompareToWorks_SPI_1609: function () {
-                // #1609
-                Bridge.Test.NUnit.Assert.True(Bridge.compare(Bridge.cast(Bridge.box(System.DateTime.fromTicks(System.Int64(0)), System.DateTime, System.DateTime.format), System.IComparable$1(System.DateTime)), System.DateTime.fromTicks(System.Int64(0)), false, System.DateTime) === 0);
-                Bridge.Test.NUnit.Assert.True(Bridge.compare(Bridge.cast(Bridge.box(System.DateTime.fromTicks(System.Int64(10000)), System.DateTime, System.DateTime.format), System.IComparable$1(System.DateTime)), System.DateTime.fromTicks(System.Int64(0)), false, System.DateTime) > 0);
-                Bridge.Test.NUnit.Assert.True(Bridge.compare(Bridge.cast(Bridge.box(System.DateTime.fromTicks(System.Int64(0)), System.DateTime, System.DateTime.format), System.IComparable$1(System.DateTime)), System.DateTime.fromTicks(System.Int64(10000)), false, System.DateTime) < 0);
-            },
-            DatePropertyWorks: function () {
-                var dt = new Date(2012, 8 - 1, 12, 13, 14, 15, 16);
-                Bridge.Test.NUnit.Assert.AreEqual(new Date(2012, 8 - 1, 12), new Date(dt.getFullYear(), dt.getMonth(), dt.getDate()));
-            },
-            DayPropertyWorks: function () {
-                var dt = new Date(2011, 7 - 1, 12, 13, 42, 56, 345);
-                Bridge.Test.NUnit.Assert.AreEqual(12, dt.getDate());
-            },
-            DayOfWeekPropertyWorks: function () {
-                var dt = new Date(2011, 8 - 1, 12, 13, 42, 56, 345);
-                Bridge.Test.NUnit.Assert.AreEqual(System.DayOfWeek.Friday, dt.getDay());
-            },
-            DayOfYearPropertyWorks: function () {
-                var dt = new Date(2011, 7 - 1, 12, 13, 42, 56, 345);
-                Bridge.Test.NUnit.Assert.AreEqual(193, Math.ceil((dt - new Date(dt.getFullYear(), 0, 1)) / 864e5));
-            },
-            HourPropertyWorks: function () {
-                var dt = new Date(2011, 7 - 1, 12, 13, 42, 56, 345);
-                Bridge.Test.NUnit.Assert.AreEqual(13, dt.getHours());
-            },
-            MillisecondPropertyWorks: function () {
-                var dt = new Date(2011, 7 - 1, 12, 13, 42, 56, 345);
-                Bridge.Test.NUnit.Assert.AreEqual(345, dt.getMilliseconds());
-            },
-            MinutePropertyWorks: function () {
-                var dt = new Date(2011, 7 - 1, 12, 13, 42, 56, 345);
-                Bridge.Test.NUnit.Assert.AreEqual(42, dt.getMinutes());
-            },
-            MonthPropertyWorks: function () {
-                var dt = new Date(2011, 7 - 1, 12, 13, 42, 56, 345);
-                Bridge.Test.NUnit.Assert.AreEqual(7, (dt.getMonth() + 1));
-            },
-            SecondPropertyWorks: function () {
-                var dt = new Date(2011, 7 - 1, 12, 13, 42, 56, 345);
-                Bridge.Test.NUnit.Assert.AreEqual(56, dt.getSeconds());
-            },
-            YearPropertyWorks: function () {
-                var dt = new Date(2011, 7 - 1, 12, 13, 42, 56, 345);
-                Bridge.Test.NUnit.Assert.AreEqual(2011, dt.getFullYear());
-            },
-            AddDaysWorks: function () {
-                var dt = new Date(2011, 7 - 1, 12, 2, 42, 56, 345);
-                var actual = new Date(dt.valueOf() + Math.round((2.5) * 864e5));
-                Bridge.Test.NUnit.Assert.AreEqual(new Date(2011, 7 - 1, 14, 14, 42, 56, 345), actual);
-                Bridge.Test.NUnit.Assert.AreEqual(new Date(2011, 7 - 1, 12, 2, 42, 56, 345), dt);
-            },
-            AddHoursWorks: function () {
-                var dt = new Date(2011, 7 - 1, 12, 2, 42, 56, 345);
-                var actual = new Date(dt.valueOf() + Math.round((2.5) * 36e5));
-                Bridge.Test.NUnit.Assert.AreEqual(new Date(2011, 7 - 1, 12, 5, 12, 56, 345), actual);
-                Bridge.Test.NUnit.Assert.AreEqual(new Date(2011, 7 - 1, 12, 2, 42, 56, 345), dt);
-            },
-            AddMillisecondsWorks: function () {
-                var dt = new Date(2011, 7 - 1, 12, 2, 42, 56, 345);
-                var actual = new Date(dt.valueOf() + Math.round(250.4));
-                Bridge.Test.NUnit.Assert.AreEqual(new Date(2011, 7 - 1, 12, 2, 42, 56, 595), actual);
-                Bridge.Test.NUnit.Assert.AreEqual(new Date(2011, 7 - 1, 12, 2, 42, 56, 345), dt);
-            },
-            AddMinutesWorks: function () {
-                var dt = new Date(2011, 7 - 1, 12, 2, 42, 56, 345);
-                var actual = new Date(dt.valueOf() + Math.round((2.5) * 6e4));
-                Bridge.Test.NUnit.Assert.AreEqual(new Date(2011, 7 - 1, 12, 2, 45, 26, 345), actual);
-                Bridge.Test.NUnit.Assert.AreEqual(new Date(2011, 7 - 1, 12, 2, 42, 56, 345), dt);
-            },
-            AddMonthsWorks: function () {
-                var dt = new Date(2011, 7 - 1, 12, 2, 42, 56, 345);
-                var actual = new Date(dt.getFullYear(), dt.getMonth() + 6, dt.getDate(), dt.getHours(), dt.getMinutes(), dt.getSeconds(), dt.getMilliseconds());
-                Bridge.Test.NUnit.Assert.AreEqual(new Date(2012, 1 - 1, 12, 2, 42, 56, 345), actual);
-                Bridge.Test.NUnit.Assert.AreEqual(new Date(2011, 7 - 1, 12, 2, 42, 56, 345), dt);
-            },
-            AddSecondsWorks: function () {
-                var dt = new Date(2011, 7 - 1, 12, 2, 42, 56, 345);
-                var actual = new Date(dt.valueOf() + Math.round((2.5) * 1e3));
-                Bridge.Test.NUnit.Assert.AreEqual(new Date(2011, 7 - 1, 12, 2, 42, 58, 845), actual);
-                Bridge.Test.NUnit.Assert.AreEqual(new Date(2011, 7 - 1, 12, 2, 42, 56, 345), dt);
-            },
-            AddYearsWorks: function () {
-                var dt = new Date(2011, 7 - 1, 12, 2, 42, 56, 345);
-                var actual = new Date(dt.getFullYear() + 3, dt.getMonth(), dt.getDate(), dt.getHours(), dt.getMinutes(), dt.getSeconds(), dt.getMilliseconds());
-                Bridge.Test.NUnit.Assert.AreEqual(new Date(2014, 7 - 1, 12, 2, 42, 56, 345), actual);
-                Bridge.Test.NUnit.Assert.AreEqual(new Date(2011, 7 - 1, 12, 2, 42, 56, 345), dt);
-            },
-            DaysInMonthWorks: function () {
-                Bridge.Test.NUnit.Assert.AreEqual(31, (new Date(2013, 1, -1).getDate() + 1));
-                Bridge.Test.NUnit.Assert.AreEqual(28, (new Date(2013, 2, -1).getDate() + 1));
-                Bridge.Test.NUnit.Assert.AreEqual(31, (new Date(2013, 3, -1).getDate() + 1));
-                Bridge.Test.NUnit.Assert.AreEqual(30, (new Date(2013, 4, -1).getDate() + 1));
-                Bridge.Test.NUnit.Assert.AreEqual(31, (new Date(2013, 5, -1).getDate() + 1));
-                Bridge.Test.NUnit.Assert.AreEqual(30, (new Date(2013, 6, -1).getDate() + 1));
-                Bridge.Test.NUnit.Assert.AreEqual(31, (new Date(2013, 7, -1).getDate() + 1));
-                Bridge.Test.NUnit.Assert.AreEqual(31, (new Date(2013, 8, -1).getDate() + 1));
-                Bridge.Test.NUnit.Assert.AreEqual(30, (new Date(2013, 9, -1).getDate() + 1));
-                Bridge.Test.NUnit.Assert.AreEqual(31, (new Date(2013, 10, -1).getDate() + 1));
-                Bridge.Test.NUnit.Assert.AreEqual(30, (new Date(2013, 11, -1).getDate() + 1));
-                Bridge.Test.NUnit.Assert.AreEqual(31, (new Date(2013, 12, -1).getDate() + 1));
-                Bridge.Test.NUnit.Assert.AreEqual(28, (new Date(2003, 2, -1).getDate() + 1));
-                Bridge.Test.NUnit.Assert.AreEqual(29, (new Date(2004, 2, -1).getDate() + 1));
-            },
-            IsLeapYearWorks: function () {
-                Bridge.Test.NUnit.Assert.True((new Date(2004, 2, -1).getDate() === 28));
-                Bridge.Test.NUnit.Assert.True((new Date(2000, 2, -1).getDate() === 28));
-                Bridge.Test.NUnit.Assert.False((new Date(2003, 2, -1).getDate() === 28));
-            }
-        }
-    });
-
-    Bridge.ns("Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests", $asm.$);
-
-    Bridge.apply($asm.$.Bridge.ClientTest.Batch4.SimpleTypes.DateTimeTests, {
-        f1: function () {
-            var dt = System.DateTime.parseExact("X", "yyyy-dd-MM");
-        },
-        f2: function () {
-            var dt = System.DateTime.parseExact("X", "yyyy-dd-MM", System.Globalization.CultureInfo.invariantCulture);
-        },
-        f3: function () {
-            var dt = System.DateTime.parseExact("X", "yyyy-dd-MM", null, true);
-        },
-        f4: function () {
-            var dt = System.DateTime.parseExact("X", "yyyy-dd-MM", System.Globalization.CultureInfo.invariantCulture, true);
-        }
-    });
-
     Bridge.define("Bridge.ClientTest.Batch4.SimpleTypes.DecimalTests", {
         methods: {
             TypePropertiesAreCorrect_SPI_1717: function () {
